requirements.txt
argcomplete==3.5.3
bitarray==3.0.0
bitstring==4.3.0
cffi==1.17.1
cryptography==44.0.1
ecdsa==0.19.0
esptool==4.8.1
intelhex==2.3.0
pycparser==2.22
pyserial==3.5
PyYAML==6.0.2
reedsolo==1.7.0
six==1.17.0
